+++ pytest_doctestplus/plugin.py
@@ -157,6 +157,20 @@ def get_optionflags(parent):
     return flag_int
 
 
+def _is_numpy_ufunc(method):
+    try:
+        import numpy as np
+    except ModuleNotFoundError:
+        # If Numpy is not installed, then there can't be any ufuncs!
+        return False
+    while True:
+        try:
+            method = method.__wrapped__
+        except AttributeError:
+            break
+    return isinstance(method, np.ufunc)
+
+
 def pytest_configure(config):
     doctest_plugin = config.pluginmanager.getplugin('doctest')
     run_regular_doctest = config.option.doctestmodules and not config.option.doctest_plus
@@ -238,7 +252,13 @@ def pytest_configure(config):
             runner = doctest.DebugRunner(
                 verbose=False, optionflags=options, checker=OutputChecker())
 
-            for test in finder.find(module):
+            tests = finder.find(module)
+            for method in module.__dict__.values():
+                if _is_numpy_ufunc(method):
+                    found = finder.find(method, module=module)
+                    tests += found
+
+            for test in tests:
                 if test.examples:  # skip empty doctests
                     ignore_warnings_context_needed = False
                     show_warnings_context_needed = False